.history/requirements_20250321184630.txt
streamlit>=1.26.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
folium>=0.14.0
requests>=2.31.0
scikit-learn>=1.3.0
altair>=5.0.0
pillow>=10.0.0
geopy>=2.3.0
hdbscan>=0.8.29
selenium>=4.1.0
webdriver-manager>=3.8.5
# Optional but recommended for enhanced functionality
# geopandas>=0.13.0
# shapely>=2.0.0